requirements.txt
# Python Backend Requirements for Lovelace

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Image Processing
Pillow>=10.1.0
rembg>=2.0.50  # Background removal
opencv-python>=4.8.1.78

# 3D Processing
trimesh>=4.0.5
pygltflib>=1.16.1

# API Clients
requests>=2.31.0
aiohttp>=3.9.0
replicate>=0.15.0  # For TripoSR and other models

# Database
sqlalchemy>=2.0.23
alembic>=1.12.1

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# Google Calendar Integration
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.108.0

# Firebase Integration
firebase-admin>=6.3.0

# Google Gemini API (for AI features and clothes search)
google-genai>=0.2.0
google-generativeai>=0.4.0

# AI/ML
openai>=1.3.0  # For AI features
anthropic>=0.7.0  # Alternative LLM

# Payment & Commerce
stripe>=8.0.0  # Stripe payment processing

# Browser Automation
playwright>=1.40.0  # Headless browser for non-UCP merchants

# Virtual Try-On (placeholder for specific libraries)
# Will depend on chosen VTO solution

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2

# Utilities
pydantic>=2.5.0
python-dateutil>=2.8.2
pytz>=2023.3
